+++ components/server/src/main/java/ome/smuggler/run/ImportYmlGen.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import ome.smuggler.config.data.ImportYmlFile;
 import ome.smuggler.config.items.ImportConfig;
-import util.config.YamlConverter;
+import util.serialization.yaml.YamlConverter;
 
 /**
  * Run this class redirecting {@code stdout} to {@code config/import.yml} to
